refactor!: avoid returning the user data from the server for edit
If somehow an attacker were to gain the session of another user, then they could
get the address of theirs. Instead just validate the session to see if it is
field and send back a boolean.

For allowing editing the address an idea would be to store the address details
in browser and fill it from there.